remove edit button on picture preview
